refactor(controller): split hotkeys definitions
