+++ src/egui_app/controller/hotkeys/actions.rs
@@ -1,225 +1,8 @@
+use super::types::{HotkeyAction, HotkeyCommand, HotkeyGesture, HotkeyScope, KeyPress};
 use crate::egui_app::state::FocusContext;
 use egui::Key;
 
-/// Identifies the surface that owns a hotkey action.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum HotkeyScope {
-    Global,
-    Focus(FocusContext),
-}
-
-impl HotkeyScope {
-    pub(crate) fn matches(self, focus: FocusContext) -> bool {
-        match self {
-            HotkeyScope::Global => true,
-            HotkeyScope::Focus(target) => target == focus,
-        }
-    }
-}
-
-/// Keyboard gesture used to trigger an action.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) struct HotkeyGesture {
-    pub(crate) first: KeyPress,
-    pub(crate) chord: Option<KeyPress>,
-}
-
-/// A single keypress plus modifier state.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) struct KeyPress {
-    pub(crate) key: Key,
-    pub(crate) command: bool,
-    pub(crate) shift: bool,
-    pub(crate) alt: bool,
-}
-
-impl HotkeyGesture {
-    pub const fn new(key: Key) -> Self {
-        Self {
-            first: KeyPress::new(key),
-            chord: None,
-        }
-    }
-
-    pub const fn with_command(key: Key) -> Self {
-        Self {
-            first: KeyPress::with_command(key),
-            chord: None,
-        }
-    }
-
-    pub const fn with_shift(key: Key) -> Self {
-        Self {
-            first: KeyPress::with_shift(key),
-            chord: None,
-        }
-    }
-
-    pub const fn with_chord(first: KeyPress, second: KeyPress) -> Self {
-        Self {
-            first,
-            chord: Some(second),
-        }
-    }
-}
-
-impl KeyPress {
-    pub const fn new(key: Key) -> Self {
-        Self {
-            key,
-            command: false,
-            shift: false,
-            alt: false,
-        }
-    }
-
-    pub const fn with_command(key: Key) -> Self {
-        Self {
-            key,
-            command: true,
-            shift: false,
-            alt: false,
-        }
-    }
-
-    pub const fn with_shift(key: Key) -> Self {
-        Self {
-            key,
-            command: false,
-            shift: true,
-            alt: false,
-        }
-    }
-
-    pub const fn with_alt(key: Key) -> Self {
-        Self {
-            key,
-            command: false,
-            shift: false,
-            alt: true,
-        }
-    }
-}
-
-/// Render a keypress in a user-friendly format (e.g. "Ctrl + G").
-pub(crate) fn format_keypress(press: &KeyPress) -> String {
-    let mut parts: Vec<&'static str> = Vec::new();
-    if press.command {
-        parts.push(command_label());
-    }
-    if press.shift {
-        parts.push("Shift");
-    }
-    if press.alt {
-        parts.push("Alt");
-    }
-    parts.push(key_label(press.key));
-    parts.join(" + ")
-}
-
-fn command_label() -> &'static str {
-    if cfg!(target_os = "macos") {
-        "Cmd"
-    } else {
-        "Ctrl"
-    }
-}
-
-fn key_label(key: Key) -> &'static str {
-    match key {
-        egui::Key::X => "X",
-        egui::Key::N => "N",
-        egui::Key::D => "D",
-        egui::Key::C => "C",
-        egui::Key::R => "R",
-        egui::Key::T => "T",
-        egui::Key::U => "U",
-        egui::Key::Y => "Y",
-        egui::Key::Z => "Z",
-        egui::Key::M => "M",
-        egui::Key::Slash => "/",
-        egui::Key::Backslash => "\\",
-        egui::Key::Quote => "'",
-        egui::Key::G => "G",
-        egui::Key::S => "S",
-        egui::Key::W => "W",
-        egui::Key::L => "L",
-        egui::Key::P => "P",
-        egui::Key::F => "F",
-        egui::Key::OpenBracket => "[",
-        egui::Key::CloseBracket => "]",
-        egui::Key::ArrowLeft => "Left",
-        egui::Key::ArrowRight => "Right",
-        egui::Key::ArrowUp => "Up",
-        egui::Key::ArrowDown => "Down",
-        _ => "Key",
-    }
-}
-
-/// Logical identifier for controller-dispatched hotkey commands.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) enum HotkeyCommand {
-    Undo,
-    Redo,
-    ToggleFocusedSelection,
-    ToggleFolderSelection,
-    NormalizeFocusedSample,
-    DeleteFocusedSample,
-    DeleteFocusedFolder,
-    RenameFocusedFolder,
-    RenameFocusedSample,
-    CreateFolder,
-    FocusFolderSearch,
-    FocusBrowserSearch,
-    AddFocusedToCollection,
-    ToggleOverlay,
-    ToggleLoop,
-    FocusWaveform,
-    FocusBrowserSamples,
-    FocusCollectionSamples,
-    FocusSourcesList,
-    FocusCollectionsList,
-    PlayRandomSample,
-    PlayPreviousRandomSample,
-    ToggleRandomNavigationMode,
-    MoveTrashedToFolder,
-    TagKeepSelected,
-    TagNeutralSelected,
-    TagTrashSelected,
-    TrimSelection,
-    FadeSelectionLeftToRight,
-    FadeSelectionRightToLeft,
-    MuteSelection,
-    NormalizeWaveform,
-    CropSelection,
-    CropSelectionNewSample,
-}
-
-/// Hotkey metadata surfaced to the UI.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub(crate) struct HotkeyAction {
-    pub(crate) id: &'static str,
-    pub(crate) label: &'static str,
-    pub(crate) gesture: HotkeyGesture,
-    scope: HotkeyScope,
-    command: HotkeyCommand,
-}
-
-impl HotkeyAction {
-    pub(crate) fn is_active(&self, focus: FocusContext) -> bool {
-        self.scope.matches(focus)
-    }
-
-    pub(crate) fn is_global(&self) -> bool {
-        matches!(self.scope, HotkeyScope::Global)
-    }
-
-    pub(crate) fn command(&self) -> HotkeyCommand {
-        self.command
-    }
-}
-
-const HOTKEY_ACTIONS: &[HotkeyAction] = &[
+pub(super) const HOTKEY_ACTIONS: &[HotkeyAction] = &[
     HotkeyAction {
         id: "undo-ctrl-z",
         label: "Undo",
@@ -509,26 +292,3 @@ const HOTKEY_ACTIONS: &[HotkeyAction] = &[
     },
 ];
 
-pub(crate) fn iter_actions() -> impl Iterator<Item = HotkeyAction> {
-    HOTKEY_ACTIONS.iter().copied()
-}
-
-pub(crate) fn focused_actions(focus: FocusContext) -> Vec<HotkeyAction> {
-    let focus = match focus {
-        FocusContext::None => FocusContext::SampleBrowser,
-        other => other,
-    };
-    HOTKEY_ACTIONS
-        .iter()
-        .copied()
-        .filter(|action| matches!(action.scope, HotkeyScope::Focus(_)) && action.is_active(focus))
-        .collect()
-}
-
-pub(crate) fn global_actions() -> Vec<HotkeyAction> {
-    HOTKEY_ACTIONS
-        .iter()
-        .copied()
-        .filter(|action| matches!(action.scope, HotkeyScope::Global))
-        .collect()
-}
